# docs/swagger.yaml
definitions:
...
  blockatlas.CollectionPage:
    items:
      properties:
        address:
          type: string
        category_address:
          type: string
        coin:
          type: integer
        description:
          type: string
        external_link:
          type: string
        id:
          type: string
        image_url:
          type: string
        name:
          type: string
        nft_version:
          type: string
        slug:
          type: string
        symbol:
          type: string
        total:
          type: integer
        type:
          type: string
      type: object
    type: array
...
  blockatlas.TxOutput:
    properties:
      address:
        type: string
      value:
        type: string
    type: object
  blockatlas.TxPage:
    items:
      properties:
        block:
          description: Height of the block the transaction was included in
          type: integer
        coin:
          description: SLIP-44 coin index of the platform
          type: integer
        date:
          description: Unix timestamp of the block the transaction was included in
          type: integer
        direction:
          description: Transaction Direction
          type: string
        error:
          description: |-
            Empty if the transaction was successful,
            else error explaining why the transaction failed (optional)
          type: string
        fee:
          description: Transaction fee (native currency)
          type: string
        from:
          description: Address of the transaction sender
          type: string
        id:
          description: Unique identifier
          type: string
        inputs:
          description: Input addresses
          items:
            $ref: '#/definitions/blockatlas.TxOutput'
          type: array
        memo:
          description: Meta data object
          type: string
        metadata:
          type: object
        outputs:
          description: Output addresses
          items:
            $ref: '#/definitions/blockatlas.TxOutput'
          type: array
        sequence:
          description: Transaction nonce or sequence
          type: integer
        status:
          description: Status of the transaction
          type: string
        to:
          description: Address of the transaction recipient
          type: string
        type:
          description: Type of metadata
          type: string
      type: object
    type: array
...
  ginutils.ApiError:
    properties:
      status_code:
        type: integer
      status_message:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /v1/{coin}/xpub/{xpub}:
    get:
      consumes:
      - application/json
      description: Get transactions from xpub address
      operationId: xpub
      parameters:
      - default: bitcoin
        description: the coin name
        in: path
        name: coin
        required: true
        type: string
      - default: zpub6ruK9k6YGm8BRHWvTiQcrEPnFkuRDJhR7mPYzV2LDvjpLa5CuGgrhCYVZjMGcLcFqv9b2WvsFtY2Gb3xq8NVq8qhk9veozrA2W9QaWtihrC
        description: the xpub address
        in: path
        name: xpub
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.TxPage'
      summary: Get xpub transactions
      tags:
      - platform
      - tx
...
  /v2/{coin}/collections/{address}:
    get:
      consumes:
      - application/json
      description: Get all collections from the address
      operationId: collections
      parameters:
      - default: ethereum
        description: the coin name
        in: path
        name: coin
        required: true
        type: string
      - default: 0x5574Cd97432cEd0D7Caf58ac3c4fEDB2061C98fB
        description: the query address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.CollectionPage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutils.ApiError'
      summary: Get Collections
      tags:
      - platform
      - collection
  /v2/{coin}/collections/{owner}/collection/{collection_id}:
    get:
      consumes:
      - application/json
      description: Get a collection from the address
      operationId: collection
      parameters:
      - default: ethereum
        description: the coin name
        in: path
        name: coin
        required: true
        type: string
      - default: 0x0875BCab22dE3d02402bc38aEe4104e1239374a7
        description: the query address
        in: path
        name: owner
        required: true
        type: string
      - default: 0x06012c8cf97bead5deae237070f9587f8e7a266d
        description: the query collection
        in: path
        name: collection_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blockatlas.CollectionPage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ginutils.ApiError'
      summary: Get Collection
      tags:
      - platform
      - collection
...
swagger: "2.0"